Fixed recent review comments
A change has been made in the IOCTL count (from 23 to 19); 4 of the ioctls i.e., VIDIOC_S_PARM, VIDIOC_QUERYCTRL, VIDIOC_S_CTRL and VIDIOC_S_EXT_CTRLS are amlogic-specific so removed them; we had kept it as we had the test cases for them and they were present in ioctl_wrapper function in westeros; but when cross verifying in other platforms they can cause issue.
They will be optional and will be included in further revisions of the documentation.